pyproject.toml
[tool.poetry]
name = "neuronpedia"
version = "1.0.2"
description = "Official Neuronpedia Python Library"
authors = ["Johnny Lin <johnny@neuronpedia.org>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
pytest = "^7.4.3"
huggingface-hub = "^0.26.2"
pandas = "^2.2.3"
tiny-dashboard = "^0.2.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120